+++ CHANGELOG.md
@@ -13,3 +13,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add RDA8910 chip flasher (Air724)
 - Fix ESP32C3,ESP32S3 flasher error
 - Add QS100 chip flasher
+- Add ESP32C3-USB chip flasher